Add tooltip for the item's name
functions.php: Initiate the varible $itemNameTooltip with the value of the $itemName because the length of its value will shrink (display of the short name). In the div "item-name" include a span element "full-item-name-tooltip" which contain the variable $itemNameTooltip.
style.css: add the class "full-item-name-tooltip" which important thing its the visibility which is hidden and the width which is fit-content. Then add some color, border-radius, padding and shadow. After that position the tooltip at the bottom of the item-name and with z-index of 1 (above of the whole grid-item). To display the tooltip change the visibility property of the "full-item-name-tooltip" to visible when the cursor is over the item-name using the hover event.