Fix some variable naming for simplicity
Remove the variable $itemNameTooltip and use the variable $itemName because it's no longer needed. The reason this happens is that we do not set the shortened item name to the variable $itemName anymore for display but in a new variable "$displayedItemName". In this way i can use the existing variable "$itemName" for the tooltip. This whole process happened because I was using the variable "$itemName" after it was truncated in the tooltip and the full item name is not displayed.