Add notification toast message
The notification bar message works exactly like the login and the register one but change some minor things to make it work in the home.php.
In more detail the changes that've been made are:
home.php: add the id="body" to the body tag and encapsulate all the functions get_toast_message() for each information using the corresponding div elements with their names. The different thing is the id for the toast which is "homeInfoToastId". Includes the info font-awesome icon and the span tag for the information text. At the bottom add the usual javascript code (function home_information_toast()) to load the toast but with the "homeInfoToastId" this time as parameter because this is the id for the whole toast.
functions.js: add the function home_information_toast() which is exactly the same as the functions "login_toast()" and "register_error_toast()" but with the only difference of the class name from which gets the text ("home-information-toast-message").
style.css: add the class name "home-content" before each appearance of the "home-information-toast". All the other necessary classes for the div's is identical to that of the login (successful logout) toast and the only thing that changed is the color for a blue-ish look and the display time for 1 second.