Display items that are images
Since I have the path of the item i can display it with a "img" tag.
functions.php: add an if-else statement to display/echo an img tag to display the image itself using the item path (if the item type is 'image') else display/echo the already existing div "icon" with the corresponding font-awesome icon.
The img tag has a built-in css style using "max-width" and "max-height" properties for the image to be displayed inside the boundaries of the grid-item.