style the information font-awesome icon
style.css: add the class "fa-info-circle" with the same properties as the "fa-check-circle" for login/style.css with the difference in color which is the same as the color of "home-information-toast-message".
Also i moved the classes "delete-item" and "fa-trash-alt" referring to "item-header" because it was in between classes referring to "item-information".