Add trash/recycle bin option (only filesystem part)
Add new file "move_to_trash.php" for handling the trash/recycle bin functionality using the move_to_trash() function. This moves the item itself in the folder "trash" in file system and changes the DB about that. At this moment works on filesystem only.
move_to_trash.php: This script it's the same as the "delete_item.php" but instead of using the "delete_item_from_filesystem()" and "delete_item_from_db()" functions uses the "move_to_trash()" function. This function (except for the item path) requires the username and that's why get it from the global variable "$_SESSION". The second difference is if the return value of the function is true did not return the "has been deleted" messege but the "has moved to trash" messege.
functions.php: At first create the folder "trash" in create_new_user_in_filesystem() function. Move the function get_user_items() above the "display_items()".  Modify the "$deleteItemLink" to redirect to the "move_to_trash.php" with the $itemPath instead of "delete_item.php". Also add the function move_to_trash(). This function has the attributes $itemPath and $username as parameters. This function moves the item within $itemPath to the trash folder of the user and make the changes to the DB (only in the filesystem at this moment). To move the item to trash folder initiate the "$source_path" which is the "$itemPath" and the "$destination_path" which is the "users_items/$username/trash/". The variable "$username" is used to access the user folder. After that attempts to execute the rename() function (as cut and paste, rename the source_path in different location) and if is correct the function return true otherwise return false. The rename function executes as:
-use the "$source_path" as first parameter (target).
-use the "$destination_path" as second parameter (destination or new name) but append the basename of the "$source_path" using the pathinfo() function.