Move the item to trash (database part)
Add the database field "in_trash" and make changes to the necessary code.
functions.php:
Include the database field to the function upload_item_in_db(), get_user_items() and display_items().
In the display_items() get the information about "in_trash" db field and check each item if its in trash and if its true "continue" which means skip (do not display) this item.
Add the function set_in_trash() which is get the item path and based on this changes the value of "in_trash" db field according to variable "$value".
Add the function set_item_path() which is get the item path and replace it in db according to the variable "$value".
Update the function move_to_trash(). First call the function as "set_in_trash($itemPath, 1)" to change the "in_trash" db field to 1 (or true) which is corresponds to the item with path shown by the variable "$itemPath". After that i change the db field "item_path" to the corresponding path before moving the item to trash folder in filesystem which the part of the string "./user_items" remain the same, append the user folder which has the same name as variable "$username", append the folder "trash" which is the target folder to contain the deleted (moved to trash) item and at last get the basename of the item to determine the new full item path to trash. To apply this change call the function set_item_path() to replace the old item path with the new inside the trash folder.